backend/pyproject.toml
[project]
name = "rag-poc-v2-backend"
version = "0.1.0"
description = "RAG Explorer v2 - Python/LangChain/LangGraph backend"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.9",
    "langsmith>=0.7.37",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-multipart>=0.0.26",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.1",
    "sse-starlette>=3.4.1",
    "unstructured[md]>=0.22.23",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"
